+++ unit/tests/public/New-BurpSuiteSite.tests.ps1
@@ -6,8 +6,10 @@ InModuleScope $env:BHProjectName {
             $parentId = 0
             $filePath = Join-Path -Path $PSScriptRoot -ChildPath '..\mocks\recorded_login.json'
             $scopeInput = [PSCustomObject]@{
-                IncludedUrls = @("http://example.com")
-                ExcludedUrls = @()
+                StartUrls = @("http://example.com")
+                InScopeUrlPrefixes = @()
+                OutOfScopeUrlPrefixes = @()
+                ProtocolOptions = 'USE_HTTP_AND_HTTPS'
             }
             $emailRecipientInput = @(
                 [PSCustomObject]@{
@@ -41,18 +43,20 @@ InModuleScope $env:BHProjectName {
             }
 
             # act
-            New-BurpSuiteSite -Name $name -ParentId $parentId -Scope $scopeInput -ScanConfigurationIds $scanConfigurationIds -EmailRecipients $emailRecipientInput -LoginCredentials $loginCredentialsInput -RecordedLogins $recordedLoginsInput -Confirm:$false
+            New-BurpSuiteSite -Name $name -ParentId $parentId -ScopeV2 $scopeInput -ScanConfigurationIds $scanConfigurationIds -EmailRecipients $emailRecipientInput -LoginCredentials $loginCredentialsInput -RecordedLogins $recordedLoginsInput -Confirm:$false
 
             # assert
             Should -Invoke _callAPI -ParameterFilter {
                 $Request.OperationName -eq "CreateSite" `
-                    -and $Request.Query -eq 'mutation CreateSite($input:CreateSiteInput!) { create_site(input:$input) { site { id name parent_id scope { included_urls excluded_urls } scan_configurations { id } application_logins { login_credentials { id label username } recorded_logins { id label } } ephemeral email_recipients { id email } } } }' `
+                    -and $Request.Query -eq 'mutation CreateSite($input:CreateSiteInput!) { create_site(input:$input) { site { id name parent_id scope_v2 { start_urls in_scope_url_prefixes out_of_scope_url_prefixes protocol_options } scan_configurations { id } application_logins { login_credentials { id label username } recorded_logins { id label } } ephemeral email_recipients { id email } } } }' `
                     -and $Request.Variables.Input.name -eq $name `
                     -and $Request.Variables.Input.parent_id -eq $parentId `
                     -and ($Request.Variables.Input.scan_configuration_ids -join ',') -eq ($scanConfigurationIds -join ',') `
                     -and $Request.Variables.Input.email_recipients[0].email -eq $emailRecipientInput[0].email `
-                    -and ($Request.Variables.Input.scope.included_urls -join ',') -eq ($scopeInput.IncludedUrls -join ',') `
-                    -and ($Request.Variables.Input.scope.excluded_urls -join ',') -eq ($scopeInput.ExcludedUrls -join ',') `
+                    -and ($Request.Variables.Input.scope_v2.start_urls -join ',') -eq ($scopeInput.StartUrls -join ',') `
+                    -and ($Request.Variables.Input.scope_v2.in_scope_url_prefixes -join ',') -eq ($scopeInput.InScopeUrlPrefixes -join ',') `
+                    -and ($Request.Variables.Input.scope_v2.out_of_scope_url_prefixes -join ',') -eq ($scopeInput.OutOfScopeUrlPrefixes -join ',') `
+                    -and $Request.Variables.Input.scope_v2.protocol_options -eq $scopeInput.ProtocolOptions `
                     -and $Request.Variables.Input.application_logins.login_credentials[0].label -eq $loginCredentialsInput[0].label `
                     -and $Request.Variables.Input.application_logins.login_credentials[0].username -eq (($loginCredentialsInput[0].Credential).GetNetworkCredential()).username `
                     -and $Request.Variables.Input.application_logins.login_credentials[0].password -eq (($loginCredentialsInput[0].Credential).GetNetworkCredential()).password `